RF_04.000.03 | Folder > delete class Folder4Test
